Continued Episode 2 from Brackeys Tutorial
If I right-click the sphere, the 'focus' field from the Player will change and show the name of the item that you right-clicked on. I'm not getting the warning message in the unity console saying that PlayerMorotr has the Vector3 as '0', which used to be pritned like 60 times persecond. I'm not getting that error message for the time being.