Player stops when Close to Sphere
When the player right clicks the Interactable Sphere, the player will move towards the sphere, but he will stop when he's close to the sphere. This fixes the bug in which the player was 'swallowing' the sphere, which could cause the 'vector3 is now 0' warning messages that came out 60 times per second in my unity console.