DEBUGGING: Console Prints 'Interacting' when Player touches sphere
When the player gets clsoe to the sphere (without touching it nor swallowing it), a debug message saying 'Interacting' gets printed 60 times per second. Fortunately, the 'Vector 3 is 0' warning message is not being printed right now. And I think it was the fact that the cause of the bug was that I decreased the interaction Radius from 3 to 1.2. Now that I icnreased it back to 3, I think I may not get the 'Vector 3 is 0' bug again.