I With Sphere Without 'Vector3' Warning Message
I can finally interact just once with the sphere when the player touches it, and only 1 message will be printed in the ocnsole per interaction. I don't have the Warning Message in the console telling me that the Vector3 was 0. That is, 'the bug is now fixed', so to speak.