I Can Rotate Cube By Dragging Mouse
I created a New Empty Scene called 'AtlasShrugged', in which I simply rendered a cube. If I click and drag on the cube, I will rotate the cube. Now, I will create a basic prism / book on Blender, and put the title 'Atlas Shrugged' on the front, and 'Published on 1957' at the back.