I can go from Anthem to Main Scene
If I press the Esc key in the Anthem scene (which is where the 'Anthem' book is located), I will go back to the Main scene, which is where the player is. I created a script to do that. Now, I will re-use that same script in the remaining 3 books.