DEBUG: I can go from Main to We the Living Scene
If I pick up any of the items, and use any of my items from the inventory, I will be sent to the We the Living scene. However, I notice a BUG in which, if I go back to the Main scene, all the picked up objects reappear on the floor, and all the items that I had on my inventory disappear.