Change Scenes When Interacting with Puzzle Wall
I created a wall, which is where the 4-book puzzle will show up if you interact with said wall. And, if you right-click on that wall, you will change to a new scene, which is currently empty. In this new scene, I will create the puzzle and ask the player to start solving the puzzle.