Attached 'We The living' Title to an Item
I attached the title 'We The Living' to one of the red potions, so that, when you drag around the potion, the title of the book will move along the potion. Also, this makes it really easy to read the title of the book, while still showing an sprite for the book.